Habilitando Flyway

+++ src/main/resources/db/migration/V2__create-table-vendas.sql
@@ -1,7 +1,7 @@
 create table vendas (
 
-    id SERIAL PRIMARY KEY,
-    cliente_id INTEGER not null,
+    id BIGINT PRIMARY KEY,
+    cliente_id BIGINT not null,
     email DATE NOT NULL
 
 );
