+++ src/main/resources/db/migration/V3__create-table-patos.sql
@@ -1,6 +1,6 @@
 create table patos (
 
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     nome VARCHAR(100) not null,
     raca VARCHAR(100)
